Separate the bot controls from the walmart specific functionality.  Create a run_bot script that will eventually be passed parameters for which product/store to hit.
